Add callsign and QTH options.

+++ src/Settings.h
@@ -91,11 +91,11 @@ namespace rose {
                     return T{r.get<int>(0), r.get<int>(1)};
                 }
                 return nullopt;
-            } else if constexpr (is_base_of_v<std::array<float,2>, T>) {
+            } else if constexpr (is_base_of_v<std::array<double,2>, T>) {
                 soci::row r;
                 sql << "SELECT a,b FROM " << real_pair_table << " WHERE name = \"" << name << '"', soci::into(r);
                 if (sql.got_data() && r.get_indicator(0) == soci::i_ok && r.get_indicator(1) == soci::i_ok) {
-                    return T{r.get<int>(0), r.get<int>(1)};
+                    return T{r.get<double>(0), r.get<double>(1)};
                 }
                 return nullopt;
             } else if constexpr (is_same_v<T, string> || is_same_v<T, const string> ||
@@ -156,7 +156,7 @@ namespace rose {
                 sql << "INSERT OR REPLACE INTO " << real_table << " (name,value) VALUES (\"" << name << "\"," << value << ')';
             } else if constexpr (is_base_of_v<std::array<int,2>,T>) {
                 sql << "INSERT OR REPLACE INTO " << int_pair_table << " (name,a,b) VALUES (\"" << name << "\"," << value.at(0) << ',' << value.at(1) << ')';
-            } else if constexpr (is_base_of_v<std::array<float,2>,T>) {
+            } else if constexpr (is_base_of_v<std::array<double,2>,T>) {
                 sql << "INSERT OR REPLACE INTO " << real_pair_table << " (name,a,b) VALUES (\"" << name << "\"," << value.at(0) << ',' << value.at(1) << ')';
             } else if constexpr (is_same_v<T,string> || is_same_v<T,const string> ||
                     is_same_v<T,string_view> || is_same_v<T,const string_view> ||
